Support for multiple concurrent active providers.
If one elects to go with GPS, there are few compelling reasons not to spin up th tower provider as well.  The power consumption is mostly from the CPU and the GPS.  There is a bit more network traffic with the tower provider, but it helps get a fix in many situations (such as indoors).